requirements.txt
langchain
langchain-openai
langchain-community
openai
python-dotenv
pypdf
faiss-cpu
langchain-text-splitters